requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.0.1
colorama==0.4.6
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.12.0
frozenlist==1.3.3
future==0.18.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
Jinja2==3.1.2
llvmlite==0.40.0
MarkupSafe==2.1.2
more-itertools==9.1.0
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numba==0.57.0
numpy==1.24.3
openai==0.27.0
openai-whisper==20230314
pydub==0.25.1
python-dotenv==0.21.1
python-telegram-bot==20.1
regex==2023.5.5
requests==2.28.2
rfc3986==1.5.0
sniffio==1.3.0
sympy==1.12
tiktoken==0.3.1
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
yarl==1.8.2
